krapper_model: extract the pure parse-output model into a standalone module (#44 brick 1b)
The model/ + model/type/ hierarchy (WrappedTU, WrappedElement, WrappedType and
friends) moves to a new :krapper_model Kotlin module (host native + jvm,
mirroring :krapper_gen) that :krapper_gen consumes via api(project). The module
builds standalone with no libclang cinterop — the physical front-end seam.

The resolve() entanglement (polymorphic overrides referencing ResolveContext /
resolvedmodel / codegen) moves to ModelResolution.kt in :krapper_gen: a
when-dispatch WrappedElement.resolve(context) extension replacing the virtual
dispatch, plus typed extensions per element so call sites keep both their
spelling and their static return types. WrappedClass's resolve-time model
mutation (modifyMethodsIfNeeded / dropConstOverloadDuplicates) and
ClassMetadata.toResolved move with it.

MethodType/AllocationStyle (the two resolved-schema enums carried as DATA on
the model) move to :krapper_model's commonMain, same package. WrappedKotlinType
stays in :krapper_gen (Kotlin naming is generator policy, not parse output);
WrappedType.kotlinType becomes an extension there. GenerationContext (pure,
holds the type intern cache) moves with the model.
