+++ krapper_model/src/nativeMain/kotlin/com/monkopedia/krapper/generator/model/type/WrappedTemplateRef.kt
