+++ settings.gradle.kts
@@ -13,6 +13,7 @@ includeBuild("compiler")
 include(":testlib_kotlin")
 include(":testlib_kotlin_manual")
 include(":krapper_gen")
+include(":krapper_model")
 include(":plugin")
 include(":slice")
 include(":feature-tests")
